This merge replaces the item selection, undo commands and
create/clone/move/remove QActions with c++14 templates and lambda
functions.

The new code enforces consistency among the various GUI subsystems and
will ensure that any changes (or future bugfixes) to the undo/selection
subsystem will affect the entire editor.

The new middleware includes View classes that provide
create/clone/move/remove actions. As most of the QActions are now
handled by these View classes the remaining MS/SI Actions were moved to
their selective Dock classes.

+++ src/gui-qt/metasprite/metasprite/palettesmodel.h
@@ -39,10 +39,10 @@ public:
     virtual QVariant data(const QModelIndex& index, int role) const final;
 
 private slots:
-    void onPaletteChanged(unsigned index);
-    void onPaletteAdded(unsigned index);
-    void onPaletteAboutToBeRemoved(unsigned index);
-    void onPaletteMoved(unsigned fromIndex, unsigned toIndex);
+    void onPaletteChanged(size_t index);
+    void onPaletteAdded(size_t index);
+    void onPaletteAboutToBeRemoved(size_t index);
+    void onPaletteMoved(size_t fromIndex, size_t toIndex);
 
 private:
     void updateAllPixmaps();
